pyproject.toml
[build-system]
requires = ["setuptools >= 65.5.0"]
build-backend = "setuptools.build_meta"

[project]
name = "322"
dynamic = ["version"]
requires-python = ">= 3.11"
authors = [ # put your name, email here!
    {name = "Gavin Travis", email = "gavin.travis@wsu.edu"},
    {name = "Dylan Kopitzke", email = "dylan.kopitzke@wsu.edu"},
    {name = "Travis Bolster", email = "travis.bolster@wsu.edu"},
]
dependencies = [
    "pymongo >= 4.0.0",
    "pyspark == 3.5.1", # latest version compatible with mongodb connector
    "graphframes-py == 0.10.0",
    "pandas",
    "matplotlib",
    "setuptools", # needed for legacy 'diskutils' during pyspark -> pandas conversion
    "Pillow", #used for tkniter images
    "numpy", #also images
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "ruff",
    "black",
    "isort",
    "pyright"
]

[project.scripts]
load_data = "scripts.load_data.load:main"
load_data_mongo = "scripts.load_data.mongo_loader:main"
analyze_links = "scripts.analysis.analyze_links:main"
graph_filter = "scripts.analysis.graph_filter:main"
trending_predictor = "scripts.analysis.trending_predictor:main"
correlation_analysis = "big_data.scripts.analysis.correlation_analysis:entrypoint"
gui = "scripts.GUI:main"

[tool.setuptools.packages.find]
where = ["big_data"]

[tool.ruff]
line-length = 120 # max line length
lint.select = [
    "E", # code style errs
    "W", # code style warnings
    "F", # pyflakes issues -> fancy set of rules
    "D", # doc string rules
    "N", # pep8 naming conventions
    "B", # bugbear errs -> more fancy rules
]
lint.ignore = [
    "D203", # incompatible with D211
    "D213", # incompatible with D212
]
exclude = [".venv", "*.md", ".egg*", ".git", ".gitignore"] # stuff not to lint + format

[tool.black]
line-length = 120 # match ruff

[tool.pyright]
venvPath = "."
venv = ".venv/"
